requirements.txt
python-telegram-bot[ext]>=20.0
google-generativeai
python-dotenv
coloredlogs
pydantic
pymongo
openai
pdfkit

pytest
pytest-asyncio
pytest-mock